restructured TUI component
