+++ dfox-tui/src/ui/mod.rs
@@ -1,0 +1,5 @@
+mod components;
+mod handlers;
+mod screens;
+
+pub use components::DatabaseClientUI;
